chrome/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Deprecated==1.2.14
executing==2.0.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
ipython==8.25.0
jedi==0.19.1
matplotlib-inline==0.1.7
mss==9.0.1
nodriver==0.37
outcome==1.3.0.post0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.46
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
stack-data==0.6.3
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==1.26.19
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
